Revert "0.0.3 - Bug fixe and removed unwanted functionalitis  "
