--- pysql_manager/core/api.py
@@ -1,1 +1,0 @@
-from pysql_manager.core.pysql import PySql
