+++ pysql_manager/functions.py
@@ -3,17 +3,14 @@
 class _ColumnFunc:
     def __init__(self, column_name):
         self.column_name = column_name
-        self._alias_name = None
+        self._alias_name = column_name
 
     def alias(self, new_column_name):
         self._alias_name = new_column_name
         return self
 
     def alias_name(self, func):
-        if not self._alias_name:
-            return f"{self.column_name}_{func}"
-
-        return f"{self._alias_name}"
+        return f"{self._alias_name}_{func}"
 
 
 def col(column_name: str):
